requirements.txt
Pillow==9.0.0
RPi.GPIO==0.7.0
inky==1.3.1

#spidev